+++ lib/active_interaction.rb
@@ -11,6 +11,7 @@ end
 
 require 'active_interaction/version'
 require 'active_interaction/errors'
+require 'active_interaction/inputs'
 
 require 'active_interaction/concerns/active_modelable'
 require 'active_interaction/concerns/active_recordable'